Split MainWindowUiTests and RequestEditorViewModelTests into focused classes
MainWindowUiTests.cs (2,430 lines, 47 tests) becomes MainWindowLayoutUiTests,
MainWindowCollectionsUiTests, MainWindowScheduledJobsUiTests,
MainWindowRequestExecutionUiTests, and MainWindowOptionsUiTests, with shared
private helpers promoted to an internal UiTestHelpers class.

RequestEditorViewModelTests.cs (738 lines, 53 tests) becomes
RequestEditorQueryParametersTests, RequestEditorHeadersTests,
RequestEditorBodyFormattingTests, and RequestEditorDraftPersistenceTests,
with the shared CreateEditor factory moved to RequestEditorTestHelpers.

No test bodies changed; all 861 solution tests still pass. Completes the
test split plan in docs/architecture/mainwindowviewmodel-split-plan.md.
