+++ docs/architecture/clean-feature-separation.md
@@ -45,6 +45,8 @@ This document answers the architecture questions raised in [issue #33](https://g
 > **Update 2026-06-14 (later 5/6):** Continuing the same pattern: moved the existing-tab lookup for a collection request into `RequestTabsWorkflow.FindMatchingTab`, the inherited-header property-change merge into `CollectionInheritedHeadersWorkflow.ObservePropertyChanges`, the `ApplySelectedCollection` projection loops into `CollectionFilterWorkflow.BuildCollectionItems`/`CollectionInheritedHeadersWorkflow.BuildHeaderViewModels`, and — most recently — the "is this an HTTP request, and which method to apply" decision from `ApplyCollectionRequestToEditor` into a new `Method` field on `CollectionRequestEditorProjectionWorkflow.BuildProjection`'s result. `MainWindowViewModel` is about 2,567 lines.
 >
 > **Status check (2026-06-14, later 6):** The Workflow/Coordinator pure-helper extraction backlog (`mainwindowviewmodel-split-plan.md` items 1–6) is now done or partial-by-design; further rounds find only marginal moves like the one above. The two substantive items remaining for "new feature requires touching only one folder" are (a) **`DockFactory` feature registrations** (step 7/2) — found to require migrating AXAML binding paths across four "thin proxy" dockable VMs (`LeftPanelViewModel`, `LogPanelViewModel`, `RequestViewModel`, `LayoutManagementViewModel`), each of which exposes only `App` (the full `MainWindowViewModel`) — a multi-PR UI-binding migration needing visual regression testing that can't be done safely headless; and (b) the **test split plan** (`MainWindowUiTests.cs` ~2,430 lines, `RequestEditorViewModelTests.cs` ~738 lines, not yet split into focused per-area classes) — lower-risk and `dotnet test`-verifiable. See the "Remaining work beyond this plan" section of `mainwindowviewmodel-split-plan.md` for details.
+>
+> **Update 2026-06-14 (later 7):** Completed the test split plan from (b) above. `MainWindowUiTests.cs` (2,430 lines, 47 tests) is now `MainWindowLayoutUiTests`, `MainWindowCollectionsUiTests`, `MainWindowScheduledJobsUiTests`, `MainWindowRequestExecutionUiTests`, and `MainWindowOptionsUiTests`, plus a shared `UiTestHelpers` internal static class for the helpers (`WaitForUiThreadAsync`, `VerifyTabRealized`, `FindDockById<T>`, `RedirectTestServer`, `AsyncStubHttpMessageHandler`) used across them. `RequestEditorViewModelTests.cs` (738 lines, 53 tests) is now `RequestEditorQueryParametersTests`, `RequestEditorHeadersTests`, `RequestEditorBodyFormattingTests`, and `RequestEditorDraftPersistenceTests`, plus a shared `RequestEditorTestHelpers` class for `CreateEditor`. No test bodies changed; `dotnet test Arbor.HttpClient.slnx` still reports 861 tests, all passing. The only item remaining from the split-plan backlog is `DockFactory` feature registrations (item 7 / (a) above).
 
 - **Monolithic main view model**: `MainWindowViewModel` (~2,500 lines at the time of writing) owns request editing, response rendering, history, collections, environments, options, scheduling, layout persistence, and logging. All UI actions pass through this single type, so responsibilities are tightly coupled.
 - **Child view models are thin proxies**: dockable VMs such as `RequestViewModel`, `ResponseViewModel`, `LeftPanelViewModel`, and `OptionsViewModel` simply forward to `MainWindowViewModel`. Reusing them elsewhere still drags the entire main VM with it.
