--- src/Arbor.HttpClient.Desktop.E2E.Tests/RequestEditorViewModelTests.cs
@@ -1,738 +1,0 @@
-using Arbor.HttpClient.Desktop.Features.HttpRequest;
-using Arbor.HttpClient.Core.Environments;
-using Arbor.HttpClient.Core.Variables;
-
-namespace Arbor.HttpClient.Desktop.E2E.Tests;
-
-/// <summary>
-/// Focused unit tests for <see cref="RequestEditorViewModel"/>.
-/// These tests do NOT require the Avalonia headless session — they exercise
-/// the request-editor logic directly (URL/query-param sync, auth header building,
-/// content-type resolution, request preview, and resolved HTTP request building).
-/// </summary>
-public class RequestEditorViewModelTests
-{
-    private static RequestEditorViewModel CreateEditor(
-        IReadOnlyList<EnvironmentVariable>? variables = null,
-        string? defaultContentType = null)
-    {
-        var variableList = variables ?? [];
-        var editor = new RequestEditorViewModel(
-            new VariableResolver(),
-            () => variableList);
-
-        if (defaultContentType is { } contentType)
-        {
-            editor.DefaultContentType = contentType;
-        }
-
-        return editor;
-    }
-
-    // ── URL ↔ query-parameter sync ──────────────────────────────────────────
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void QueryParameters_ArePopulated_WhenUrlContainsQueryString()
-    {
-        var editor = CreateEditor();
-        editor.RequestUrl = "http://localhost:5000/search?q=hello&page=2";
-
-        editor.RequestQueryParameters.Should().HaveCount(3);
-        editor.RequestQueryParameters[0].Key.Should().Be("q");
-        editor.RequestQueryParameters[0].Value.Should().Be("hello");
-        editor.RequestQueryParameters[1].Key.Should().Be("page");
-        editor.RequestQueryParameters[1].Value.Should().Be("2");
-        editor.RequestQueryParameters[2].Key.Should().BeEmpty();
-        editor.RequestQueryParameters[2].IsEnabled.Should().BeFalse();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void QueryParameters_AreCleared_WhenUrlHasNoQuery()
-    {
-        var editor = CreateEditor();
-        editor.RequestUrl = "http://localhost:5000/search?q=hello";
-        editor.RequestUrl = "http://localhost:5000/search";
-
-        editor.RequestQueryParameters.Should().HaveCount(1);
-        editor.RequestQueryParameters[0].Key.Should().BeEmpty();
-        editor.RequestQueryParameters[0].IsEnabled.Should().BeFalse();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void Url_IsUpdated_WhenQueryParameterValueChanges()
-    {
-        var editor = CreateEditor();
-        editor.RequestUrl = "http://localhost:5000/items?foo=bar";
-
-        editor.RequestQueryParameters[0].Value = "baz";
-
-        editor.RequestUrl.Should().Contain("foo=baz");
-        editor.RequestUrl.Should().NotContain("foo=bar");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void Url_IsUpdated_WhenDisabledQueryParameterIsExcluded()
-    {
-        var editor = CreateEditor();
-        editor.RequestUrl = "http://localhost:5000/items?a=1&b=2";
-
-        editor.RequestQueryParameters[0].IsEnabled = false;
-
-        editor.RequestUrl.Should().NotContain("a=1");
-        editor.RequestUrl.Should().Contain("b=2");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void AddQueryParameter_AddsEmptyEntryToCollection()
-    {
-        var editor = CreateEditor();
-        editor.RequestUrl = "http://localhost:5000/items";
-        editor.AddQueryParameterCommand.Execute().Subscribe();
-
-        editor.RequestQueryParameters.Should().HaveCount(1);
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void RemoveQueryParameter_RemovesEntry()
-    {
-        var editor = CreateEditor();
-        editor.RequestUrl = "http://localhost:5000/items?x=1";
-        var param = editor.RequestQueryParameters[0];
-        editor.RemoveQueryParameterCommand.Execute(param).Subscribe();
-
-        editor.RequestQueryParameters.Should().HaveCount(1);
-        editor.RequestQueryParameters[0].Key.Should().BeEmpty();
-        editor.RequestQueryParameters[0].IsEnabled.Should().BeFalse();
-    }
-
-    // ── Request headers ─────────────────────────────────────────────────────
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void AddHeader_AddsEmptyHeader()
-    {
-        var editor = CreateEditor();
-        editor.AddHeaderCommand.Execute().Subscribe();
-
-        editor.RequestHeaders.Should().HaveCount(1);
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void RemoveHeader_RemovesHeader()
-    {
-        var editor = CreateEditor();
-        editor.AddHeaderCommand.Execute().Subscribe();
-        var header = editor.RequestHeaders[0];
-        editor.RemoveHeaderCommand.Execute(header).Subscribe();
-
-        editor.RequestHeaders.Should().HaveCount(1);
-        editor.RequestHeaders[0].Name.Should().BeEmpty();
-        editor.RequestHeaders[0].IsEnabled.Should().BeFalse();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void PlaceholderHeader_AutoEnablesAndAppendsNewPlaceholder_WhenNameIsTyped()
-    {
-        var editor = CreateEditor();
-        var placeholder = editor.RequestHeaders[0];
-        placeholder.Name.Should().BeEmpty();
-        placeholder.IsEnabled.Should().BeFalse();
-
-        placeholder.Name = "X-Custom";
-
-        placeholder.IsEnabled.Should().BeTrue();
-        editor.RequestHeaders.Should().HaveCount(2);
-        editor.RequestHeaders[^1].Name.Should().BeEmpty();
-        editor.RequestHeaders[^1].IsEnabled.Should().BeFalse();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void PlaceholderHeader_CannotBeEnabled_WhenNameIsBlank()
-    {
-        var editor = CreateEditor();
-        var placeholder = editor.RequestHeaders[0];
-
-        placeholder.IsEnabled = true;
-
-        editor.RequestHeaders.Should().HaveCount(1);
-        editor.RequestHeaders[0].Name.Should().BeEmpty();
-        editor.RequestHeaders[0].IsEnabled.Should().BeFalse();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void PlaceholderQueryParameter_AutoEnablesAndAppendsNewPlaceholder_WhenKeyIsTyped()
-    {
-        var editor = CreateEditor();
-        var placeholder = editor.RequestQueryParameters[0];
-        placeholder.Key.Should().BeEmpty();
-        placeholder.IsEnabled.Should().BeFalse();
-
-        placeholder.Key = "page";
-
-        placeholder.IsEnabled.Should().BeTrue();
-        editor.RequestQueryParameters.Should().HaveCount(2);
-        editor.RequestQueryParameters[^1].Key.Should().BeEmpty();
-        editor.RequestQueryParameters[^1].IsEnabled.Should().BeFalse();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void PlaceholderQueryParameter_CannotBeEnabled_WhenKeyIsBlank()
-    {
-        var editor = CreateEditor();
-        var placeholder = editor.RequestQueryParameters[0];
-
-        placeholder.IsEnabled = true;
-
-        editor.RequestQueryParameters.Should().HaveCount(1);
-        editor.RequestQueryParameters[0].Key.Should().BeEmpty();
-        editor.RequestQueryParameters[0].IsEnabled.Should().BeFalse();
-    }
-
-    // ── Auth header building ─────────────────────────────────────────────────
-
-    [AvaloniaTheory(Timeout = 10_000)]
-    [InlineData(RequestEditorViewModel.AuthNoneOption, "Authorization: Bearer token")]
-    public void RequestPreview_DoesNotContainAuthHeader_WhenAuthModeIsNone(string mode, string notExpected)
-    {
-        var editor = CreateEditor();
-        editor.SelectedAuthModeOption = mode;
-        editor.AuthBearerToken = "token";
-
-        editor.RequestPreview.Should().NotContain(notExpected);
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void RequestPreview_ContainsBearerToken_WhenBearerAuthModeSelected()
-    {
-        var editor = CreateEditor();
-        editor.SelectedAuthModeOption = RequestEditorViewModel.AuthBearerOption;
-        editor.AuthBearerToken = "my-token";
-
-        editor.RequestPreview.Should().Contain("Authorization: Bearer my-token");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void RequestPreview_ContainsBasicCredentials_WhenBasicAuthModeSelected()
-    {
-        var editor = CreateEditor();
-        editor.SelectedAuthModeOption = RequestEditorViewModel.AuthBasicOption;
-        editor.AuthBasicUsername = "user";
-        editor.AuthBasicPassword = "pass";
-
-        // Base64 of "user:pass"
-        var expected = Convert.ToBase64String(System.Text.Encoding.UTF8.GetBytes("user:pass"));
-        editor.RequestPreview.Should().Contain($"Authorization: Basic {expected}");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void RequestPreview_BearerOverridesExistingAuthorizationHeader()
-    {
-        var editor = CreateEditor();
-        editor.SelectedAuthModeOption = RequestEditorViewModel.AuthBearerOption;
-        editor.AuthBearerToken = "new-token";
-        editor.AddHeaderCommand.Execute().Subscribe();
-        editor.RequestHeaders[0].Name = "Authorization";
-        editor.RequestHeaders[0].Value = "Bearer old-token";
-        editor.RequestHeaders[0].IsEnabled = true;
-
-        editor.RequestPreview.Should().Contain("Authorization: Bearer new-token");
-        editor.RequestPreview.Should().NotContain("Bearer old-token");
-    }
-
-    // ── Content-type resolution ──────────────────────────────────────────────
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void ContentType_IsEmpty_WhenNoneSelected()
-    {
-        var editor = CreateEditor();
-        editor.SelectedContentTypeOption = RequestEditorViewModel.NoneContentTypeOption;
-
-        editor.ContentType.Should().BeEmpty();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void ContentType_IsCustomValue_WhenCustomOptionSelectedAndValueProvided()
-    {
-        var editor = CreateEditor();
-        editor.SelectedContentTypeOption = RequestEditorViewModel.CustomContentTypeOption;
-        editor.CustomContentType = "application/vnd.api+json";
-
-        editor.ContentType.Should().Be("application/vnd.api+json");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void RequestPreview_ContainsDefaultContentType_WhenBodyNonEmptyAndNoExplicitType()
-    {
-        var editor = CreateEditor(defaultContentType: "application/json");
-        editor.SelectedContentTypeOption = RequestEditorViewModel.NoneContentTypeOption;
-        editor.RequestBody = "{\"key\":\"value\"}";
-
-        editor.RequestPreview.Should().Contain("Content-Type: application/json");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void RequestPreview_DoesNotContainContentType_WhenBodyIsEmpty()
-    {
-        var editor = CreateEditor(defaultContentType: "application/json");
-        editor.RequestBody = string.Empty;
-
-        editor.RequestPreview.Should().NotContain("Content-Type:");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void RequestPreview_PrettyPrintsResolvedBody_WhenEnabled()
-    {
-        var editor = CreateEditor();
-        editor.SelectedContentTypeOption = "application/json";
-        editor.RequestBody = "{\"a\":1}";
-        editor.PrettyPrintRequestBody = true;
-        editor.PrettyPrintRequestBodyUseIndentation = true;
-
-        editor.RequestPreview.Should().Contain("{\n  \"a\": 1\n}");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void ShowRequestPreview_IsEnabledByDefault()
-    {
-        var editor = CreateEditor();
-
-        editor.ShowRequestPreview.Should().BeTrue();
-        editor.IsRequestPreviewPanelVisible.Should().BeTrue();
-    }
-
-    // ── Variable resolution in preview ───────────────────────────────────────
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void RequestPreview_ResolvesVariables_UsingActiveEnvironment()
-    {
-        var variables = new List<EnvironmentVariable>
-        {
-            new("host", "localhost:5000", IsEnabled: true),
-            new("token", "abc123", IsEnabled: true)
-        };
-        var editor = CreateEditor(variables);
-        editor.RequestUrl = "http://{{host}}/items";
-        editor.SelectedAuthModeOption = RequestEditorViewModel.AuthBearerOption;
-        editor.AuthBearerToken = "{{token}}";
-
-        editor.RequestPreview.Should().Contain("http://localhost:5000/items");
-        editor.RequestPreview.Should().Contain("Authorization: Bearer abc123");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void RequestPreview_ResolvesUnknownVariablesToEmpty_WhenVariablesNotDefined()
-    {
-        var editor = CreateEditor();
-        editor.RequestUrl = "https://{{host}}/items";
-
-        // VariableResolver replaces unresolved tokens with an empty string.
-        editor.RequestPreview.Should().Contain("https:///items");
-    }
-
-    // ── BuildResolvedHttpRequestDraft ────────────────────────────────────────
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_ReturnsCorrectMethodAndUrl()
-    {
-        var editor = CreateEditor();
-        editor.SelectedMethod = "POST";
-        editor.RequestUrl = "http://localhost:5000/users";
-        editor.RequestName = "Create user";
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.Method.Should().Be("POST");
-        draft.Url.Should().Be("http://localhost:5000/users");
-        draft.Name.Should().Be("Create user");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_ResolvesVariables_InUrlAndBody()
-    {
-        var variables = new List<EnvironmentVariable>
-        {
-            new("base", "http://localhost:5000", IsEnabled: true)
-        };
-        var editor = CreateEditor(variables);
-        editor.RequestUrl = "{{base}}/items";
-        editor.RequestBody = "{\"url\":\"{{base}}\"}";
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.Url.Should().Be("http://localhost:5000/items");
-        draft.Body.Should().Be("{\"url\":\"http://localhost:5000\"}");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_PrettyPrintsJsonBodyWithIndentation_WhenEnabled()
-    {
-        var editor = CreateEditor();
-        editor.SelectedContentTypeOption = "application/json";
-        editor.RequestBody = "{\"a\":1}";
-        editor.PrettyPrintRequestBody = true;
-        editor.PrettyPrintRequestBodyUseIndentation = true;
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.Body.Should().Be("{\n  \"a\": 1\n}");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_PrettyPrintsJsonBodyWithoutIndentation_WhenEnabled()
-    {
-        var editor = CreateEditor();
-        editor.SelectedContentTypeOption = "application/json";
-        editor.RequestBody = "{  \"a\" : 1  }";
-        editor.PrettyPrintRequestBody = true;
-        editor.PrettyPrintRequestBodyUseIndentation = false;
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.Body.Should().Be("{\"a\":1}");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_PrettyPrintJson_DoesNotEscapeAngleBracketsOrAmpersands()
-    {
-        var editor = CreateEditor();
-        editor.SelectedContentTypeOption = "application/json";
-        editor.RequestBody = "{\"html\":\"<div>&</div>\"}";
-        editor.PrettyPrintRequestBody = true;
-        editor.PrettyPrintRequestBodyUseIndentation = false;
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.Body.Should().Be("{\"html\":\"<div>&</div>\"}");
-        draft.Body.Should().NotContain("\\u003c");
-        draft.Body.Should().NotContain("\\u003e");
-        draft.Body.Should().NotContain("\\u0026");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_PrettyPrintsXmlBodyWithoutIndentation_WhenEnabled()
-    {
-        var editor = CreateEditor();
-        editor.SelectedContentTypeOption = "application/xml";
-        editor.RequestBody = "<root>\n  <item>1</item>\n</root>";
-        editor.PrettyPrintRequestBody = true;
-        editor.PrettyPrintRequestBodyUseIndentation = false;
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.Body.Should().Be("<root><item>1</item></root>");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void PrettyPrintRequestBodySourceCommand_FormatsSourceBody()
-    {
-        var editor = CreateEditor();
-        editor.SelectedContentTypeOption = "application/json";
-        editor.PrettyPrintRequestBodyUseIndentation = false;
-        editor.RequestBody = "{ \"a\" : 1 }";
-
-        editor.PrettyPrintRequestBodySourceCommand.Execute().Subscribe();
-
-        editor.RequestBody.Should().Be("{\"a\":1}");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_UsesFollowRedirectsForRequest()
-    {
-        var editor = CreateEditor();
-        editor.FollowRedirectsForRequest = false;
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.FollowRedirects.Should().BeFalse();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void ValidateUrlBeforeSend_IsEnabledByDefault()
-    {
-        var editor = CreateEditor();
-
-        editor.ValidateUrlBeforeSend.Should().BeTrue();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_SetsIgnoreCertificateValidation_WhenEnabled()
-    {
-        var editor = CreateEditor();
-        editor.IgnoreCertificateValidationForRequest = true;
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.IgnoreCertificateValidation.Should().BeTrue();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_LeavesIgnoreCertificateValidationNull_WhenDisabled()
-    {
-        var editor = CreateEditor();
-        editor.IgnoreCertificateValidationForRequest = false;
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.IgnoreCertificateValidation.Should().BeNull();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_UsesNullTlsVersionOverride_WhenDefaultOptionSelected()
-    {
-        var editor = CreateEditor();
-        editor.SelectedTlsVersionOverrideOption = RequestEditorViewModel.DefaultTlsVersionOverrideOption;
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.TlsVersionOverride.Should().BeNull();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_UsesPerRequestTlsVersionOverride_WhenSpecified()
-    {
-        var editor = CreateEditor();
-        editor.SelectedTlsVersionOverrideOption = "Tls13";
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.TlsVersionOverride.Should().Be("Tls13");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_UsesPerRequestTimeoutSeconds_WhenProvided()
-    {
-        var editor = CreateEditor();
-        editor.RequestTimeoutSecondsText = "12";
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.TimeoutSeconds.Should().Be(12);
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_UsesNoTimeout_WhenPerRequestTimeoutIsZero()
-    {
-        var editor = CreateEditor();
-        editor.RequestTimeoutSecondsText = "0";
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.TimeoutSeconds.Should().Be(0);
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_UsesNullTimeout_WhenPerRequestTimeoutIsBlank()
-    {
-        var editor = CreateEditor();
-        editor.RequestTimeoutSecondsText = " ";
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.TimeoutSeconds.Should().BeNull();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_UsesNullTimeout_WhenPerRequestTimeoutContainsNoDigits()
-    {
-        var editor = CreateEditor();
-        editor.RequestTimeoutSecondsText = "abc";
-        var draft = editor.BuildResolvedHttpRequestDraft();
-        draft.TimeoutSeconds.Should().BeNull();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_ClampsPerRequestTimeoutToMaximum()
-    {
-        var editor = CreateEditor();
-        editor.RequestTimeoutSecondsText = "101";
-        var draft = editor.BuildResolvedHttpRequestDraft();
-        draft.TimeoutSeconds.Should().Be(100);
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void RequestTimeoutSecondsText_ShouldClear_WhenInputContainsNonDigits()
-    {
-        var editor = CreateEditor();
-        editor.RequestTimeoutSecondsText = "a1b2c3";
-
-        editor.RequestTimeoutSecondsText.Should().BeEmpty();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void RequestTimeoutSecondsText_ShouldClampToMaximum_WhenInputIsNumeric()
-    {
-        var editor = CreateEditor();
-        editor.RequestTimeoutSecondsText = "123";
-
-        editor.RequestTimeoutSecondsText.Should().Be("100");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_UsesSelectedHttpVersion()
-    {
-        var editor = CreateEditor();
-        editor.SelectedHttpVersionOption = "2.0";
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.HttpVersion.Should().Be(System.Net.HttpVersion.Version20);
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_IncludesSelectedContentTypeInHeaders_WhenBodyProvided()
-    {
-        var editor = CreateEditor();
-        editor.SelectedContentTypeOption = "application/xml";
-        editor.SelectedMethod = "POST";
-        editor.RequestBody = "<root/>";
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.Headers.Should().ContainSingle(h =>
-            string.Equals(h.Name, "Content-Type", StringComparison.OrdinalIgnoreCase) &&
-            h.Value == "application/xml");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BuildResolvedHttpRequestDraft_DoesNotDuplicateContentTypeHeader_WhenUserAlsoAddsOneManually()
-    {
-        var editor = CreateEditor();
-        editor.SelectedContentTypeOption = "application/xml";
-        editor.RequestBody = "<root/>";
-        editor.AddHeaderCommand.Execute().Subscribe();
-        editor.RequestHeaders[0].Name = "Content-Type";
-        editor.RequestHeaders[0].Value = "text/plain";
-        editor.RequestHeaders[0].IsEnabled = true;
-
-        var draft = editor.BuildResolvedHttpRequestDraft();
-
-        draft.Headers.Should().ContainSingle(h =>
-            string.Equals(h.Name, "Content-Type", StringComparison.OrdinalIgnoreCase));
-        var contentTypeHeader = draft.Headers!.First(h => string.Equals(h.Name, "Content-Type", StringComparison.OrdinalIgnoreCase));
-        contentTypeHeader.Value.Should().Be("application/xml");
-    }
-
-    // ── HTTP version ─────────────────────────────────────────────────────────
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void OptionsAffectingPropertyChanged_IsFired_WhenHttpVersionChanges()
-    {
-        var fired = false;
-        var editor = new RequestEditorViewModel(
-            new VariableResolver(),
-            () => [],
-            onOptionsAffectingPropertyChanged: () => fired = true);
-
-        editor.SelectedHttpVersionOption = "2.0";
-
-        fired.Should().BeTrue();
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void OptionsAffectingPropertyChanged_IsNotFired_WhenHttpVersionSetToSameValue()
-    {
-        var count = 0;
-        var editor = new RequestEditorViewModel(
-            new VariableResolver(),
-            () => [],
-            onOptionsAffectingPropertyChanged: () => count++);
-
-        editor.SelectedHttpVersionOption = "1.1"; // same as default
-
-        count.Should().Be(0);
-    }
-
-    // ── GetResolvedVariables ─────────────────────────────────────────────────
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void GetResolvedVariables_FiltersOutDisabledVariables()
-    {
-        var variables = new List<EnvironmentVariable>
-        {
-            new("a", "1", IsEnabled: true),
-            new("b", "2", IsEnabled: false)
-        };
-        var editor = CreateEditor(variables);
-
-        var resolved = editor.GetResolvedVariables();
-
-        resolved.Should().HaveCount(1);
-        resolved[0].Name.Should().Be("a");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void GetResolvedVariables_FiltersOutVariablesWithEmptyName()
-    {
-        var variables = new List<EnvironmentVariable>
-        {
-            new("", "value", IsEnabled: true),
-            new("   ", "other", IsEnabled: true)
-        };
-        var editor = CreateEditor(variables);
-
-        var resolved = editor.GetResolvedVariables();
-
-        resolved.Should().BeEmpty();
-    }
-
-    // ── BulkUpdate suppression guard ────────────────────────────────────────
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void BeginBulkUpdate_SuppressesRefreshDuringScope_ThenFiresOneRefreshOnDispose()
-    {
-        var editor = CreateEditor();
-        editor.SelectedContentTypeOption = "application/json";
-        editor.RequestUrl = "http://localhost/test";
-
-        var previewChangedCount = 0;
-        editor.PropertyChanged += (_, e) =>
-        {
-            if (e.PropertyName == nameof(RequestEditorViewModel.RequestPreview))
-            {
-                previewChangedCount++;
-            }
-        };
-
-        // Set multiple properties inside the suppression window.
-        using (editor.BeginBulkUpdate())
-        {
-            editor.SelectedMethod = "POST";
-            editor.RequestBody = "{}";
-            editor.RequestUrl = "http://localhost/bulk";
-            // No refresh should have fired yet.
-            previewChangedCount.Should().Be(0);
-        }
-
-        // Exactly one refresh should have fired when the handle was disposed.
-        previewChangedCount.Should().Be(1);
-        editor.RequestPreview.Should().Contain("POST");
-        editor.RequestPreview.Should().Contain("/bulk");
-    }
-
-    [AvaloniaFact(Timeout = 10_000)]
-    public void EndBulkUpdate_AfterBeginBulkUpdate_FiresOneRefresh()
-    {
-        var editor = CreateEditor();
-        editor.RequestUrl = "http://localhost/before";
-
-        var previewChangedCount = 0;
-        editor.PropertyChanged += (_, e) =>
-        {
-            if (e.PropertyName == nameof(RequestEditorViewModel.RequestPreview))
-            {
-                previewChangedCount++;
-            }
-        };
-
-        editor.BeginBulkUpdate();
-        editor.SelectedMethod = "PUT";
-        editor.RequestUrl = "http://localhost/after";
-        previewChangedCount.Should().Be(0);
-
-        editor.EndBulkUpdate();
-
-        previewChangedCount.Should().Be(1);
-        editor.RequestPreview.Should().Contain("PUT");
-        editor.RequestPreview.Should().Contain("/after");
-    }
-}
-
